ThumbnailGalleryWidget: Do not downsize thumbnail => more speed.
As counter intuitive as this may sound, but the overall payload is smaller, if we do not optimize this for 50% of the screen. Reason: We do need the 100% screen version anyway and it will be automatically loaded by the browser. So with this patch we save around 0.5 MB download payload.