[prerender] storePrerenderedContent now extends _redirects with uppercase urls.
This is a workaround to netlifys url normalization. If one has a permalink with an upcase such as "myPermalink", netlify would previously redirect to "/mypermalink". Once Scrivito JS is started it would no longer recognize that permalink, since permalinks are case-sensitive. With an explicit redirect entry in "_redirects", Netlify does not "normalize" the given URL to lower-case.