[webpack] No longer use file-loader for html files.
Motivation: In the near future we want to use html-webpack-plugin - but this does not work well with file-loader.

Currently the only place, where we used the file-loader for html files was in ContactFormWidget which is a Netlify only feature. I think it's fine that we use webpack inline imports (see [1]). If we every switch away from webpack we will see very early, that something needs to be done here.

[1] https://webpack.js.org/concepts/loaders/#inline