Make transform also work for copy-webpack-plugin 6.1.0.
If multiple patterns reference similar or identical files (e.g. "../public" and "../public/_headers") it is not well defined, which pattern will be executed - the first or the last? To avoid this confusion now "../public" is copied, but only "public/_headers" is transformed.

copy-webpack-plugin 6.0.1 most likely broke because of https://github.com/webpack-contrib/copy-webpack-plugin/pull/484.