Use npm 7 on Travis and Netlify
I already switch locally to npm 7 in [1]. But I did not yet switch to npm 7 on Travis and Netlify. Now is the time.

Without this change subtle failures could emerge. E.g. an update of webpack locally ran fine, but failed on Netlify.

This can be removed, as soon as the node version installs npm v7 by default. According to [2] this is node version 15 or newer.

Inspired by [3] and [4].

[1] https://github.com/Scrivito/scrivito_example_app_js/pull/375
[2] https://nodejs.org/en/download/releases/
[3] travis-ci/travis-ci#4653 (comment)
[4] https://docs.netlify.com/configure-builds/manage-dependencies/#npm