ThumbnailGallery: Replace react-bnb-gallery with slick and bootstrap.
Sadly "react-bnb-gallery" is no longer maintained (if it ever where, see [1]) and does not support React 17 or 18.

To not run into the same issue with yet-another gallery, let's pick one, that we're already using: react-slick. For it to work, it needs to be shown in a dialog, which react-bootstrap provides. Since "react-slick" is already part of the package (for testimonial) its not wrapped in a "loadable" component.

[1] https://github.com/peterpalau/react-bnb-gallery