[FormContainer] Store a hidden "url" field
During prerender it uses the URL of the containing obj. Once in a browser it will use the current location with pathname and search, but not hash.

See concept [1]

Please note, that history.location does not provide the origin and also does not provide a "url" (See [2]). Since we have to mingle the parts of the location anyway the hash can be dropped.

[1] https://docs.google.com/document/d/1dpLXc9grQkUqxeuVX63lbPwFm--htFDpWdBpi-k-txw/edit#heading=h.jokyvkkj2bgb
[2] https://github.com/remix-run/history/blob/dev/docs/api-reference.md#location